+++ contracts/os/andromeda-kernel/src/contract.rs
@@ -101,8 +101,9 @@ pub fn execute(
         ExecuteMsg::Send { message } => execute::send(execute_env, message),
         ExecuteMsg::TriggerRelay {
             packet_sequence,
-            packet_ack_msg,
-        } => execute::trigger_relay(execute_env, packet_sequence, packet_ack_msg),
+            channel_id,
+            packet_ack,
+        } => execute::trigger_relay(execute_env, packet_sequence, channel_id, packet_ack),
         ExecuteMsg::UpsertKeyAddress { key, value } => {
             execute::upsert_key_address(execute_env, key, value)
         }
@@ -173,5 +174,8 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<Binary, ContractErr
         QueryMsg::Version {} => encode_binary(&ADOContract::default().query_version(deps)?),
         QueryMsg::AdoType {} => encode_binary(&ADOContract::default().query_type(deps)?),
         QueryMsg::Owner {} => encode_binary(&ADOContract::default().query_contract_owner(deps)?),
+        QueryMsg::ChainNameByChannel { channel } => {
+            encode_binary(&query::chain_name_by_channel(deps, channel)?)
+        }
     }
 }